fix: add missing include and type casts for UI components
- Add core.h include to MockDisplay.h
- Fix uint16_t/uint32_t casts in Text.cpp for color parameters

+++ src/libs/common/ui/components/Text.cpp
@@ -47,7 +47,7 @@ namespace common::ui::components
 
         // Set text properties
         display.setTextSize(_fontSize);
-        display.setTextColor(_foregroundColor, _backgroundColor);
+        display.setTextColor(static_cast<uint16_t>(_foregroundColor), static_cast<uint16_t>(_backgroundColor));
 
         // Calculate text position based on alignment
         int textX = _x;
@@ -93,7 +93,7 @@ namespace common::ui::components
         {
             int charHeight = 8 * _fontSize;
             int underlineY = textY + charHeight;
-            display.drawLine(textX, underlineY, textX + _width, underlineY, _foregroundColor);
+            display.drawLine(textX, underlineY, textX + _width, underlineY, static_cast<uint32_t>(_foregroundColor));
         }
 
         return common::patterns::Result<void>::success();
